docker/requirements.txt
scrapy
selenium
psycopg2-binary
pandas
